Add non-empty href attributes to anchors
Travis ci had htmlproofer errors warning about missing href attributes
on the About page. Traced to Back-to-top and detector map <a> tags.

Add #top as back-to-top link in those anchors, though it only takes us
to the top of the content.

Set href in detector <map> to same value as in the data-modal attribute.
This seems to play nicely with bootstrap but may need further work.