Add and document target for updating Gems
As reported in #66, Github helpfully warns about significant issues
in Gem versions in the lock file.

To help updating Gem dependencies, add an `update-gems` target in
the makefile to automate the update and regenerate Gemfile.lock.
Do not auto-commit changes to avoid clashes.

Update README with info on running this target and committing changes.

No changes to Gemfile.lock are done yet.